split sequence impl to cpp file
